requirements.txt
# requirements.txt

rank_bm25  #-0.2.2
langchain-cohere   #-0.4.6
langchain-huggingface   #-0.3.1

sentence-transformers  #-5.1.1
torch  #-2.8.0
langchain  #-0.3.27
langchain-community  #-0.3.30
faiss-cpu  #-1.12.0
python-dotenv  #-1.1.1
langchain-google-genai  #-2.1.12
requests  #-2.3.3
beautifulsoup4  #-4.14.2 soupsieve-2.8
Flask